Adding rating info to files
I updated the data collection code to include vehicle rating information, on a scale of 0-10 (10 is the best). Ratings come from Car and Driver. The rating information could be useful if, for example, unit sales data were later added to augment the analysis. Then, a correlation could be drawn between Car and Driver's rating of a model and its total sales to see if the Car and Driver ratings reflect popular opinion (based on a vehicle class or price segment).